+++ app/src/main/java/com/pict/pbl/medswift/screens/HomeScreen.kt
@@ -6,11 +6,13 @@ import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
 import androidx.activity.viewModels
 import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.padding
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.filled.History
 import androidx.compose.material.icons.filled.Home
+import androidx.compose.material.icons.filled.LocalHospital
 import androidx.compose.material.icons.filled.Person
 import androidx.compose.material3.*
 import androidx.compose.runtime.Composable
@@ -21,6 +23,7 @@ import androidx.compose.ui.graphics.vector.ImageVector
 import androidx.compose.ui.platform.LocalConfiguration
 import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.tooling.preview.Preview
+import androidx.compose.ui.unit.dp
 import androidx.navigation.NavController
 import androidx.navigation.NavHostController
 import androidx.navigation.compose.NavHost
@@ -67,7 +70,10 @@ class HomeScreen : ComponentActivity() {
                         startActivity( this )
                     }
                 }) {
-                    Text(text = "Diagnosis")
+                    Row {
+                        Icon( Icons.Default.LocalHospital , contentDescription = "Diagnosis" )
+                        Text(text = "Diagnosis" , modifier = Modifier.padding( 8.dp ) )
+                    }
                 }
             }
         ) { padding ->