+++ app/src/main/java/com/pict/pbl/medswift/screens/RegisterScreen.kt
@@ -6,6 +6,7 @@ import androidx.activity.compose.setContent
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.layout.width
 import androidx.compose.foundation.lazy.LazyColumn
 import androidx.compose.foundation.text.KeyboardActions
 import androidx.compose.foundation.text.KeyboardOptions
@@ -18,13 +19,16 @@ import androidx.compose.runtime.*
 import androidx.compose.runtime.saveable.rememberSaveable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.focus.FocusDirection
+import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.platform.LocalFocusManager
+import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.text.input.ImeAction
 import androidx.compose.ui.text.input.KeyboardType
 import androidx.compose.ui.text.input.PasswordVisualTransformation
 import androidx.compose.ui.text.input.VisualTransformation
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
+import androidx.compose.ui.unit.sp
 import com.google.firebase.auth.ktx.auth
 import com.google.firebase.ktx.Firebase
 import com.pict.pbl.medswift.auth.CurrentUser
@@ -62,6 +66,24 @@ class RegisterScreen : ComponentActivity() {
     private fun ActivityUI() {
         // TODO: Add register screen UI here
         LazyColumn(){
+            item {
+                Text(
+                    text = "History" ,
+                    color = Color.Black ,
+                    fontWeight = FontWeight.Bold ,
+                    fontSize = 22.sp ,
+                    modifier = Modifier
+                        .padding( 24.dp )
+                        .fillMaxWidth()
+                )
+                Divider(
+                    color = MaterialTheme.colorScheme.primary,
+                    modifier = Modifier
+                        .padding( bottom = 8.dp )
+                        .fillMaxWidth()
+                        .width(1.dp)
+                )
+            }
             item{
                 UserFirstName(
                     modifier = Modifier